Revert "refactor(logging): move sentry config"
This reverts commit cf5bd41bbf445ffeb24a5e1eb1d500f5921a5f21.
