+++ packages/vip-go/config/getSentryConfig.js
@@ -12,10 +12,9 @@ const getSentryConfig = () => {
   if (!sentryConfig) {
     sentryConfig = path.join(
       __dirname,
-      '../sentry.config.js',
+      'sentry.config.js',
     );
   }
-
   return sentryConfig;
 };
