+++ packages/vip-go/config/sentry.config.js
